refactor: localize FastAPI model contracts
